smart_plant_care_system/requirements.txt
mysql-connector-python==8.0.33
customtkinter==5.2.1
Pillow==10.0.1
matplotlib==3.7.2
numpy==1.24.3